lab-03-serving-monitoring/requirements.txt
# Core MLOps
dagster
dagster-webserver
dagster-mlflow
mlflow

# ML / DL
torch
torchvision
pandas
numpy
matplotlib
Pillow

# Serving / UI
gradio

# Monitoring / Drift
evidently

# Optional (falls ihr es nutzt)
scikit-learn

openpyxl

evidently
